requirements.txt
alembic==1.14.0
apispec==6.7.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Migrate==4.0.7
flask-smorest==0.45.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
mysqlclient==2.2.5
packaging==24.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
waitress==3.0.1
webargs==8.6.0
Werkzeug==3.1.3
